requirements.txt
black==23.3.0
click==8.1.3
Flask~=2.2.2
flask_cors==3.0.10
flask_sqlalchemy==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe~=2.1.1
marshmallow==3.19.0
marshmallow_sqlalchemy==0.29.0
pip==23.1
pg8000==1.29.4
python-dotenv==1.0.0
setuptools==67.6.1
sqlalchemy==2.0.9
Werkzeug~=2.2.2
wheel~=0.37.1
PyJWT==2.6.0
gunicorn==20.1.0